ETL/requirements.txt
statsmodels
numpy
scikit-learn
prometheus_api_client
datetime
pyyaml
confluent-kafka